requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
sqlalchemy
python-multipart
flwr
